I am very disappointed in his behavior
